Fixed issues with saving power calibrations.
There is a bug in pyqtgraph.configfile such that arrays cannot be read
back in if they are long enough to contain line breaks (that are
introduced by repr(array)). To get around this we can convert arrays to
strings, and remove line breaks before sending them to configfile, then
reconvert them back to arrays.

I also changed the key for the calibration to be a string of the
wavelength instead of a float.